+++ sources/Google.Solutions.IapDesktop.Extensions.Management/ToolWindows/SerialOutput/XtermReader.cs
@@ -1,5 +1,5 @@
 ﻿//
-// Copyright 2020 Google LLC
+// Copyright 2024 Google LLC
 //
 // Licensed to the Apache Software Foundation (ASF) under one
 // or more contributor license agreements.  See the NOTICE file
@@ -21,89 +21,41 @@
 
 using Google.Solutions.Common.Text;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
+using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
-using VtNetCore.Exceptions;
-using VtNetCore.VirtualTerminal;
-using VtNetCore.XTermParser;
-using VtNetCore.XTermParser.SequenceType;
 
 namespace Google.Solutions.IapDesktop.Extensions.Management.ToolWindows.SerialOutput
 {
     /// <summary>
-    /// Reader that extracts pure text out of a string containing
-    /// ANSI escape sequences.
+    /// Reader that removes Xterm control sequences from a stream
+    /// that are commonly encountered in serial console streams. This
+    /// is done on a best-effort basis.
     /// </summary>
-    public class XtermReader : IAsyncReader<string>
+    internal class XtermReader : IAsyncReader<string>
     {
+        /// <summary>
+        /// Regex patterns for control sequences that are being sanitized.
+        /// </summary>
+        private static readonly string[] ControlSequencePatterns = new[] {
+            "\u001b\\[2J",     // Clear the screen.
+            "\u001b\\[01;01H", // Set cursor position to the top-left corner.
+            "\u001b\\[=3h",    // Set the terminal to a application keypad mode.
+        };
+
+        private static readonly Regex allControlSequencePatterns = new Regex(
+            string.Join("|", ControlSequencePatterns));
+
         private readonly IAsyncReader<string> reader;
-        private readonly VirtualTerminalController controller;
-        private readonly XTermInputBuffer buffer;
 
         public XtermReader(IAsyncReader<string> reader)
         {
             this.reader = reader;
-            this.controller = new VirtualTerminalController();
-            this.buffer = new XTermInputBuffer();
         }
 
-        private IEnumerable<TerminalSequence> ParseSequences(string rawData)
-        {
-            this.buffer.Add(Encoding.UTF8.GetBytes(rawData));
-
-            var sequences = new List<TerminalSequence>();
-            this.controller.ClearChanges();
-
-            while (!this.buffer.AtEnd)
-            {
-                try
-                {
-                    var terminalSequence = XTermSequenceReader.ConsumeNextSequence(
-                        this.buffer,
-                        this.controller.IsUtf8());
-                    if (terminalSequence.ProcessFirst != null)
-                    {
-                        sequences.AddRange(terminalSequence.ProcessFirst);
-                    }
-
-                    sequences.Add(terminalSequence);
-                }
-                catch (IndexOutOfRangeException)
-                {
-                    this.buffer.PopAllStates();
-                    break;
-                }
-                catch (ArgumentException)
-                {
-                    if (!this.buffer.AtEnd)
-                    {
-                        this.buffer.ReadRaw();
-                    }
-
-                    this.buffer.Commit();
-                }
-                catch (EscapeSequenceException)
-                {
-                    if (!this.buffer.AtEnd)
-                    {
-                        this.buffer.ReadRaw();
-                    }
-
-                    this.buffer.Commit();
-                }
-                catch (Exception ex4)
-                {
-                    throw new InvalidOperationException("Failed to process xterm input", ex4);
-                }
-            }
-
-            this.buffer.Flush();
-
-            return sequences;
-        }
+        //----------------------------------------------------------------------
+        // IAsyncReader.
+        //----------------------------------------------------------------------
 
         public async Task<string> ReadAsync(CancellationToken token)
         {
@@ -111,10 +63,7 @@ namespace Google.Solutions.IapDesktop.Extensions.Management.ToolWindows.SerialOu
                 .ReadAsync(token)
                 .ConfigureAwait(false);
 
-            return new string(ParseSequences(chunk)
-                .OfType<CharacterSequence>()
-                .Select(s => s.Character)
-                .ToArray());
+            return allControlSequencePatterns.Replace(chunk, string.Empty);
         }
 
         public void Dispose()
